+++ include/parasol/modules/font.h
@@ -56,7 +56,7 @@ struct FontList {
    struct FontList * Next;    // Pointer to the next entry in the list.
    STRING Name;               // The name of the font face.
    STRING Alias;              // Reference to another font Name if this is an alias.
-   LONG * Points;             // Pointer to an array of fixed point sizes supported by the font.
+   int *  Points;             // Pointer to an array of fixed point sizes supported by the font.
    STRING Styles;             // Supported styles are listed here in CSV format.
    STRING Axes;               // For variable fonts, lists all supported axis codes in CSV format
    BYTE   Scalable;           // TRUE if the font is scalable.
@@ -87,22 +87,22 @@ class objFont : public Object {
    struct RGB8 Underline;  // Enables font underlining when set.
    struct RGB8 Colour;     // The font colour in RGB8 format.
    FTF    Flags;           // Optional flags.
-   LONG   Gutter;          // The 'external leading' value, measured in pixels.  Applies to fixed fonts only.
-   LONG   LineSpacing;     // The amount of spacing between each line.
-   LONG   X;               // The starting horizontal position when drawing the font string.
-   LONG   Y;               // The starting vertical position when drawing the font string.
-   LONG   TabSize;         // Defines the tab size to use when drawing and manipulating a font string.
-   LONG   WrapEdge;        // Enables word wrapping at a given boundary.
-   LONG   FixedWidth;      // Forces a fixed pixel width to use for all glyphs.
-   LONG   Height;          // The point size of the font, expressed in pixels.
-   LONG   Leading;         // 'Internal leading' measured in pixels.  Applies to fixed fonts only.
-   LONG   MaxHeight;       // The maximum possible pixel height per character.
+   int    Gutter;          // The 'external leading' value, measured in pixels.  Applies to fixed fonts only.
+   int    LineSpacing;     // The amount of spacing between each line.
+   int    X;               // The starting horizontal position when drawing the font string.
+   int    Y;               // The starting vertical position when drawing the font string.
+   int    TabSize;         // Defines the tab size to use when drawing and manipulating a font string.
+   int    WrapEdge;        // Enables word wrapping at a given boundary.
+   int    FixedWidth;      // Forces a fixed pixel width to use for all glyphs.
+   int    Height;          // The point size of the font, expressed in pixels.
+   int    Leading;         // 'Internal leading' measured in pixels.  Applies to fixed fonts only.
+   int    MaxHeight;       // The maximum possible pixel height per character.
    ALIGN  Align;           // Sets the position of a font string to an abstract alignment.
-   LONG   AlignWidth;      // The width to use when aligning the font string.
-   LONG   AlignHeight;     // The height to use when aligning the font string.
-   LONG   Ascent;          // The total number of pixels above the baseline.
-   LONG   EndX;            // Indicates the final horizontal coordinate after completing a draw operation.
-   LONG   EndY;            // Indicates the final vertical coordinate after completing a draw operation.
+   int    AlignWidth;      // The width to use when aligning the font string.
+   int    AlignHeight;     // The height to use when aligning the font string.
+   int    Ascent;          // The total number of pixels above the baseline.
+   int    EndX;            // Indicates the final horizontal coordinate after completing a draw operation.
+   int    EndY;            // Indicates the final vertical coordinate after completing a draw operation.
 
    // Action stubs
 
@@ -176,49 +176,49 @@ class objFont : public Object {
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setGutter(const LONG Value) noexcept {
+   inline ERR setGutter(const int Value) noexcept {
       if (this->initialised()) return ERR::NoFieldAccess;
       this->Gutter = Value;
       return ERR::Okay;
    }
 
-   inline ERR setLineSpacing(const LONG Value) noexcept {
+   inline ERR setLineSpacing(const int Value) noexcept {
       this->LineSpacing = Value;
       return ERR::Okay;
    }
 
-   inline ERR setX(const LONG Value) noexcept {
+   inline ERR setX(const int Value) noexcept {
       this->X = Value;
       return ERR::Okay;
    }
 
-   inline ERR setY(const LONG Value) noexcept {
+   inline ERR setY(const int Value) noexcept {
       this->Y = Value;
       return ERR::Okay;
    }
 
-   inline ERR setTabSize(const LONG Value) noexcept {
+   inline ERR setTabSize(const int Value) noexcept {
       this->TabSize = Value;
       return ERR::Okay;
    }
 
-   inline ERR setWrapEdge(const LONG Value) noexcept {
+   inline ERR setWrapEdge(const int Value) noexcept {
       this->WrapEdge = Value;
       return ERR::Okay;
    }
 
-   inline ERR setFixedWidth(const LONG Value) noexcept {
+   inline ERR setFixedWidth(const int Value) noexcept {
       this->FixedWidth = Value;
       return ERR::Okay;
    }
 
-   inline ERR setHeight(const LONG Value) noexcept {
+   inline ERR setHeight(const int Value) noexcept {
       if (this->initialised()) return ERR::NoFieldAccess;
       this->Height = Value;
       return ERR::Okay;
    }
 
-   inline ERR setMaxHeight(const LONG Value) noexcept {
+   inline ERR setMaxHeight(const int Value) noexcept {
       if (this->initialised()) return ERR::NoFieldAccess;
       this->MaxHeight = Value;
       return ERR::Okay;
@@ -229,33 +229,33 @@ class objFont : public Object {
       return ERR::Okay;
    }
 
-   inline ERR setAlignWidth(const LONG Value) noexcept {
+   inline ERR setAlignWidth(const int Value) noexcept {
       this->AlignWidth = Value;
       return ERR::Okay;
    }
 
-   inline ERR setAlignHeight(const LONG Value) noexcept {
+   inline ERR setAlignHeight(const int Value) noexcept {
       this->AlignHeight = Value;
       return ERR::Okay;
    }
 
-   inline ERR setEndX(const LONG Value) noexcept {
+   inline ERR setEndX(const int Value) noexcept {
       this->EndX = Value;
       return ERR::Okay;
    }
 
-   inline ERR setEndY(const LONG Value) noexcept {
+   inline ERR setEndY(const int Value) noexcept {
       this->EndY = Value;
       return ERR::Okay;
    }
 
-   inline ERR setBold(const LONG Value) noexcept {
+   inline ERR setBold(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[20];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setItalic(const LONG Value) noexcept {
+   inline ERR setItalic(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[5];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
@@ -278,8 +278,8 @@ class objFont : public Object {
 struct FontBase {
 #ifndef PARASOL_STATIC
    ERR (*_GetList)(struct FontList **Result);
-   LONG (*_StringWidth)(objFont *Font, CSTRING String, LONG Chars);
-   LONG (*_CharWidth)(objFont *Font, ULONG Char);
+   int (*_StringWidth)(objFont *Font, CSTRING String, int Chars);
+   int (*_CharWidth)(objFont *Font, ULONG Char);
    ERR (*_RefreshFonts)(void);
    ERR (*_SelectFont)(CSTRING Name, CSTRING Style, CSTRING *Path, FMETA *Meta);
    ERR (*_ResolveFamilyName)(CSTRING String, CSTRING *Result);
@@ -291,8 +291,8 @@ struct FontBase {
 extern struct FontBase *FontBase;
 namespace fnt {
 inline ERR GetList(struct FontList **Result) { return FontBase->_GetList(Result); }
-inline LONG StringWidth(objFont *Font, CSTRING String, LONG Chars) { return FontBase->_StringWidth(Font,String,Chars); }
-inline LONG CharWidth(objFont *Font, ULONG Char) { return FontBase->_CharWidth(Font,Char); }
+inline int StringWidth(objFont *Font, CSTRING String, int Chars) { return FontBase->_StringWidth(Font,String,Chars); }
+inline int CharWidth(objFont *Font, ULONG Char) { return FontBase->_CharWidth(Font,Char); }
 inline ERR RefreshFonts(void) { return FontBase->_RefreshFonts(); }
 inline ERR SelectFont(CSTRING Name, CSTRING Style, CSTRING *Path, FMETA *Meta) { return FontBase->_SelectFont(Name,Style,Path,Meta); }
 inline ERR ResolveFamilyName(CSTRING String, CSTRING *Result) { return FontBase->_ResolveFamilyName(String,Result); }
@@ -300,8 +300,8 @@ inline ERR ResolveFamilyName(CSTRING String, CSTRING *Result) { return FontBase-
 #else
 namespace fnt {
 extern ERR GetList(struct FontList **Result);
-extern LONG StringWidth(objFont *Font, CSTRING String, LONG Chars);
-extern LONG CharWidth(objFont *Font, ULONG Char);
+extern int StringWidth(objFont *Font, CSTRING String, int Chars);
+extern int CharWidth(objFont *Font, ULONG Char);
 extern ERR RefreshFonts(void);
 extern ERR SelectFont(CSTRING Name, CSTRING Style, CSTRING *Path, FMETA *Meta);
 extern ERR ResolveFamilyName(CSTRING String, CSTRING *Result);
